Use latest mvn dependency versions (#352)
This is a routine update of all the maven dependency versions.

The junit jupiter family of dependency versions now no longer uses a pom
wide property to define its version. This is to make it easier to
automatically update these fields using the command
`mvn versions:use-latest-versions` which automates this. Using a
property disrupts this from working.

Signed-off-by: Jason Katonica <katonica@us.ibm.com>